Extract experience view component per naming scheme (migration step 5).
Move exp route UI into experience-view.js so page.js stays a thin Next.js adapter.
